feat/intelligence_ts_examples_nextjs-web-chats: Adds dockerfile + instructions and make build work
